use float

+++ apis/k8ssandra/v1alpha1/k8ssandracluster_types.go
@@ -526,9 +526,9 @@ const (
 	ServerDistributionDse       = ServerDistribution("dse")
 )
 
-func (kc *K8ssandraCluster) DefaultNumTokens(serverVersion *semver.Version) int64 {
+func (kc *K8ssandraCluster) DefaultNumTokens(serverVersion *semver.Version) float64 {
 	if kc.Spec.Cassandra.ServerType == ServerDistributionCassandra && serverVersion.Major() == 3 {
-		return int64(256)
+		return float64(256)
 	}
-	return int64(16)
+	return float64(16)
 }
